Keep authorization view within architecture budget
